+++ src/View/ExceptionStrategy.php
@@ -182,7 +182,10 @@ EOT;
                 if (is_callable($this->message)) {
                     $callback = $this->message;
                     $message = (string) $callback($exception, $this->displayExceptions);
-                } elseif ($this->displayExceptions && $exception instanceof \Exception) {
+                } elseif ($this->displayExceptions
+                    // @todo clean up once PHP 7 requirement is enforced
+                    && ($exception instanceof \Exception || $exception instanceof \Throwable)
+                ) {
                     $previous = '';
                     $previousException = $exception->getPrevious();
                     while ($previousException) {